requirements.txt
django-environ==0.11.2
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers
whitenoise
setuptools